mlx5: fix a use-after-free error in mlx5_next_poll
When using cq_ex inteface, if the user destroys the QP associated with
the current work completion, the next ibv_next_poll() call will cause a
use-after-free error since it needs to access the QP that has already
been destroyed through cq->cur_rsc inside get_req_context().

Fix this error by resetting the cq->cur_rsc in __mlx5_cq_clean if it is
associated with the QP to be destroyed.